+++ include/simdutf/scalar/utf8_to_decomposed/utf8_to_decomposed.h
@@ -163,7 +163,7 @@ size_t decompose_bmp(uint16_t code_point, char *output, uint8_t *first_ccc,
   for (size_t k = 0; k < length; k++) {
     *output++ = bytes[k];
   }
-  uint8_t ccc_delta = value >> 29;
+  uint8_t ccc_delta = uint8_t(value >> 29);
   *first_ccc = ccc_delta == 0 ? 0 : *ccc - ccc_delta;
   return output - start;
 }
@@ -359,7 +359,7 @@ bool check_code_point_bmp(uint16_t code_point, size_t *out_length,
                           uint8_t *ccc) {
   uint16_t value = lookup_check_trie<form>(code_point);
   *out_length += value & 0x3F;
-  *ccc = (value >> 6) & 0xFF;
+  *ccc = uint8_t((value >> 6) & 0xFF);
   return !(value >> 15);
 }
 
@@ -369,7 +369,7 @@ static bool check_code_point_supplementary(uint32_t code_point,
   uint64_t kv = lookup_supplementary_code_point<form>(code_point);
   uint32_t k = kv & 0x1FFFFF;
   if (k == code_point) {
-    uint8_t length = 0;
+    size_t length = 0;
     uint16_t offset = (kv >> 21) & 0xFFFF;
     uint32_t const *chars;
     if constexpr (form == DecomposedForm::NFD) {
